requirements.txt
openai==1.56.1
langchain==0.1.6
langchain-community==0.0.20
langchain-core==0.1.23
langchain-experimental==0.0.49
langchain-openai==0.0.5
pandas==2.2.2
python-dotenv==1.0.1
streamlit